ManytoOne corrigé

+++ src/main/java/com/loparok/model/Address.java
@@ -2,7 +2,6 @@ package com.loparok.model;
 
 import java.sql.Date;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -17,7 +16,7 @@ public class Address {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private int id;
 
     private String denomination;
